+++ CHANGELOG.md
@@ -83,6 +83,7 @@
 * [BUGFIX] Fix issue with orphaned group nodes in vParquet5-preview6 [#6095](https://github.com/grafana/tempo/pull/6095) (@stoewer)
 * [BUGFIX] Fix issues related to integer dedicated columns in vParquet5-preview2 [#5716](https://github.com/grafana/tempo/pull/5716) (@stoewer)
 * [BUGFIX] Fix disappearing dedicated event attrs in trace view and dedicated blob column validation in vParquet5-preview6 [#6100](https://github.com/grafana/tempo/pull/6100) (@stoewer)
+* [BUGFIX] Fix dedicated column array handling to better support attributes with mixed array and non-array values [#6199](https://github.com/grafana/tempo/pull/6199) (@mdisibio)
 * [ENHANCEMENT] Add integer attribute analysis to the tempo-cli `analyse block(s)` commands [#6103](https://github.com/grafana/tempo/pull/6103) (@mdisibio)
 
 ### Tempo Rearchitecture [EXPERIMENTAL]